frontend/requirements.txt
streamlit
requests
pandas
plotly
pillow
python-dotenv 